move all printing-related logic to internal module
