+++ src/internal.rs
@@ -1,0 +1,5 @@
+mod formatter;
+mod lines_with_endings;
+mod outputstyle;
+
+pub use outputstyle::OutputStyle;
