+++ src/lib.rs
@@ -1,5 +1,3 @@
 pub mod clash;
-mod formatter;
-pub mod outputstyle;
 pub mod solution;
 pub mod stub;
